Add files via upload
Homework 1:  Some text manipulation

 Write a program to do the following:

 Input a paragraph of at least 3 sentences, preferably more. (Copy a paragraph or two from a news article or a web page about some subject of interest. )

In our first class, we split the sentence into words by splitting on the blank space between words.  That is the default for split.  You can also split on other characters by naming them.  Look it up. 

Separate your input into sentences.  For now, just separate on the ‘.’ that ends most sentences.

 Print out the following, with an appropriate label for each and reasonable spacing:

How many sentences were there?
Be careful.  Explore a bit.  Check the answer.
What is the last sentence?  Is this a problem? 
How would you fix it?
How many total words in all the sentences? (Report the number of words found in the list.  It is not necessary to remove non-words yet.)
Print out the longest sentence (longest in words, not characters.)
Print out the shortest sentence
Sort the words in the list.
Print out the sorted list of words.
Take a look at the sorted list of words to see what the default processes consider a word.
Test your program carefully.  Realize that when it is tested, the input used will be different from yours.  Thus, you must not allow your program to be specific to the particular paragraph that you read.